+++ changelog.md
@@ -3,8 +3,9 @@
   * Rich farmland has been disabled by default
   * Crops can now grow with bone meal when farmland is not moist
 * Added back Sleep features: Cloth, Tiredness
-* Added back Death features: Death
-  * On death the player will now lose 15% of durability on the items in the inventory and will drop 30% of their inventory items 
+* Added back Death features: Death Penalty (renamed from Death), Player killer bounty (split from Death)
+  * On death the player will now lose 15% of durability on the items in the inventory and will drop 30% of their inventory items
+  * Player killer bounty now accounts for multiple unique players killed, increasing the experience multiplier
   * Removed Grave
 * Added back Item features: Unvanishable Items (aka Unbreakable Items)
 * Added back World features: Coal & Fire